Use BigDecimal for non-integers in program

+++ StockWatch.Backend/src/main/java/com/henning/oden/java/StockAlert/Backend/dto/StockWatchCreationRequest.java
@@ -2,10 +2,12 @@ package com.henning.oden.java.StockAlert.Backend.dto;
 
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 public class StockWatchCreationRequest {
     private String stockCode;
-    private double minPrice;
-    private double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private int alertThreshold;
 }
